Move core.py out of _libs
